ML_System_Design/Agents/ReAct_Agent/requirements.txt
# Core dependencies for ReAct Agent
openai>=1.0.0
google-search-results>=2.4.0
python-dotenv>=1.0.0
